+++ CHANGES.md
@@ -14,6 +14,7 @@
 * removed `class Additions`, now self-contained in [GitHub Updater Additions](https://github.com/afragen/github-updater-additions)
 * refactored `Install::install()` a bit more
 * use new `github_updater_admin_pages` filter hook for adding `index.php` from Remote Management
+* ensure that all API install fields are available for all installed APIs
 
 #### 7.6.2 / 2018-04-27
 * move `auth_required` stuff from `Base` to `Settings`